refactor(app): extract headless-approve-delete from headless.ts
Moves the respond/lifecycle family (approve/reject/input/select, cancel/cancel-workflow/delete-workflow/detach-workflow/open-terminal) out of headless.ts into headless-approve-delete.ts, behavior unchanged.
One atomic slice of the step-4 large-file decomposition.
